remove class & methods related to gviz_api
The new shore logs are being incorporated into TethysDash, and including the gviz_api code here introduces an extra dependency that most people won't use. I don't think it should be in there for the public version of the tool, and if we reintroduce it in the future, we should do it in a way such that the extra dependency is only required if users specifically enable and load it.